backend/requirements.txt
# FastAPI server
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart
pydantic
pydantic-settings

# Database
sqlalchemy>=2.0
alembic

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Core AI
openai-whisper
sentence-transformers
numpy
scipy

# LangChain & LangGraph
langchain>=0.2.0
langchain-community
langchain-text-splitters
langchain-groq
langgraph>=0.1.0

# Vector DB
qdrant-client>=1.9.0

# Audio
sounddevice
ffmpeg-python
soundfile

# Tracing
langsmith

# Config
python-dotenv

# Export (PDF highlights)
reportlab
python-docx

# Testing
pytest
pytest-asyncio
httpx
